Fix for DNS Server value disappeared
:Release Notes:
Fix for DNS Server value is disappeared when
clicked on Connect Button

:Detailed Notes:
DNS address was missing even though connman has
the address internally.Nameservers property
changed call back is added so that the address is
changed when connman gets it.

:Testing Performed:
Tested Locally.

:QA Notes:
NA

:Issues Addressed:
[WRP-5183] [OSE -RPI4]: Newly entered DNS Server
value is disappeared when clicked on Connect
button

Change-Id: Ifea1b48e32704ef69b4bf876083268edc67cd4bb