Fix for [OSE-RPi4] : 'Wi-Fi Network is disconnected' popup is displayed when connecting to a WiFi Network
:Release Notes:
Fix for [OSE-RPi4] : 'Wi-Fi Network is disconnected'
popup is displayed when connecting to a WiFi Network

:Detailed Notes:
When connect command is executed to connect to a
wifi network that is already connected, the
WiFi is disconnected and disconect pop getting on the
UI So that we removed remove_connected_service method
so it will not disconnect from network.

:Testing Performed:
Tested Locally.

:QA Notes:
NA

:Issues Addressed:
[WRP-9151] [OSE-RPi4] : 'Wi-Fi Network is disconnected'
popup is displayed when connecting to a WiFi Network

Change-Id: Ic6c955825164175ef3db692bfa976046a9a471a2