Fix to enable/disable persistentMode for P2P
:Release Notes:
Declare enable_persistent_mode as static global variable

:Detailed Notes:
Passing address of local variable to a timeout function,
this is creating a invalid memory access when the lifetime
of the variable is not valid when accessed in the timeout callback.

Declared enable_persistent_mode as static global to make the variable
available when accessed after the given timeout in callback function.

:Testing Performed:
Manualy verified the Test case - QWTC-7221

:QA Notes:
N/A

:Issues Addressed:
[WRR-15353] [OSE- RPI4 ] :In terminal we are getting "persistentMode"
            as "disabled".

Change-Id: I1226295643fb9be265b58adec1a561a5ab4034b2